requirements.txt
Django==3.2.10
django-taggit==3.1.0
Pillow==11.1.0
python-dotenv==1.0.0